factor out opcode semantics
