+++ src/main/scala/translating/GTIRBToIR.scala
@@ -72,13 +72,11 @@ class TempIf(
   */
 class GTIRBToIR(
   mods: Seq[Module],
-  parserMap: Option[immutable.Map[String, List[InsnSemantics]]],
+  lifter: InsnLoader,
   cfg: CFG,
   mainAddress: Option[BigInt],
   mainName: Option[String]
 ) {
-
-  val opcodeBytes = 4
 
   private val functionNames = MapDecoder.decode_uuid(mods.map(_.auxData("functionNames").data))
   private val functionEntries = MapDecoder.decode_set(mods.map(_.auxData("functionEntries").data))
@@ -86,42 +84,6 @@ class GTIRBToIR(
 
   // maps block UUIDs to their address
   private val blockUUIDToAddress = createAddresses()
-
-  case class BlockPos(
-    b: ByteInterval.Block,
-    uuid: ByteString,
-    address: Long,
-    offset: Long,
-    size: Long,
-    byteIntervalContent: ByteString
-  ) {
-    assert(blockUUIDToAddress(uuid) == address)
-
-    def content: ByteString = {
-      byteIntervalContent.substring(offset.toInt, (offset + size).toInt)
-    }
-
-    def opcodes: Seq[Int] = Range
-      .Exclusive(0, size.toInt, opcodeBytes)
-      .map(i => content.substring(i, i + opcodeBytes))
-      .map(b => bytesToi32(b.toByteArray, true))
-
-    def toStatements(): Seq[Seq[Statement]] = {
-      offlineLifter.Lifter.liftBlockBytes(opcodes, address)
-    }
-
-    private def bytesToi32(bytes: Array[Byte], littleEndian: Boolean): Int = {
-      val buffer = java.nio.ByteBuffer.wrap(bytes)
-      if (littleEndian) {
-        buffer.order(java.nio.ByteOrder.LITTLE_ENDIAN).getInt
-      } else {
-        buffer.getInt
-      }
-    }
-
-  }
-
-  private lazy val uuidToBlockContent = createOpcodeBlocks()
 
   // mapping from a symbol's UUID to the symbol itself
   private val uuidToSymbol = mods.flatMap(_.symbols).map(s => s.uuid -> s).toMap
@@ -161,28 +123,6 @@ class GTIRBToIR(
     }).toMap
 
     blockAddresses
-  }
-
-  def createOpcodeBlocks(): immutable.Map[ByteString, BlockPos] = {
-    val blocks = mods
-      .flatMap(_.sections)
-      .flatMap(_.byteIntervals)
-      .map(bi => {
-        assert(
-          bi.hasAddress
-        ) // Unsure if this holds for PIC https://grammatech.github.io/gtirb/cpp/classgtirb_1_1_byte_interval.html#aaf13ececea7d2b943402feeb4f1aae35
-        (bi.blocks.toList, bi.contents, bi.address)
-      })
-
-    val codeblocks = blocks.flatMap((bl, cont, bi_addr) =>
-      bl.collect(b =>
-        b.value.code match {
-          case Some(c) =>
-            c.uuid -> BlockPos(b, c.uuid, bi_addr + b.offset, b.offset, c.size, cont)
-        }
-      )
-    )
-    codeblocks.toMap
   }
 
   // maps block UUIDs to their outgoing edges
@@ -242,23 +182,13 @@ class GTIRBToIR(
 
     // maybe good to sort blocks by address around here?
 
-    val semanticsLoader
-      : (blockUUID: ByteString, blockCountIn: Int, blockAddress: Option[BigInt]) => Seq[Seq[Statement]] =
-      parserMap match {
-        case Some(parserMap) => {
-          val semanticsLoader = GTIRBLoader(parserMap)
-          (uuid, blockCount, addr) => semanticsLoader.visitBlock(uuid, blockCount, addr).toSeq
-        }
-        case None => (uuid, _, _) => uuidToBlockContent(uuid).toStatements()
-      }
-
     for ((functionUUID, blockUUIDs) <- functionBlocks) {
       val procedure = uuidToProcedure(functionUUID)
       var blockCount = 0
       for (blockUUID <- blockUUIDs) {
         val block = uuidToBlock(blockUUID)
 
-        val statements = semanticsLoader(blockUUID, blockCount, block.address)
+        val statements = lifter.decodeBlock(blockUUID, blockCount, block.address)
         blockCount += 1
         for ((stmts, i) <- statements.zipWithIndex) {
           block.statements.addAll(insertPCIncrement(stmts))
